requirements.txt
absl-py
ml-collections==0.1.1
accelerate
transformers==4.28.0
tokenizers==0.13.2
diffusers
ipython
omegaconf==2.3.0
open-clip-torch
timm==0.6.13
git+https://github.com/openai/CLIP.git
torch~=2.0
torchvision
tqdm
wandb
opencv-python-headless
einops
insightface==0.7.3
onnxruntime
pyfacer